map reader and wmts cap reader should be separate
